rename costants.py to constants.py
