+++ src/aind_metadata_mapper/open_ephys/utils/constants.py
